Camera pt2, thruster adjustments, lateral thrust pt1, thruster graphics
drawShip now uses the camera. The only thing missing from the camera at
this point is zoom.

Thruster functions will now clamp the given strength parameters to
between plus or minus the given ship's thruster strength. This allows
greater flexibility in the design of the stabilizer functions, and is
just generally a good check to do.

I've written the lateral thrust function, though it isn't hooked up to
any keys and it hasn't been tested. Medial and lateral stabilizer
functions still need to be written.

Thrusters have graphics. Right now they're the same for every ship save
for color, which is ship-specific.