Active thruster model
Thruster functions will no longer set accelerations themselves. Instead,
they'll add their strength parameter to their component of the
activeThrusters object within the ship object. The updateShip function
will handle converting the active thruster components into an
acceleration vector before it increments velocity and position. An
important note here is that because activeThrusters is used for drawing,
it can't be cleared in the update function, and now that it is used for
more than just drawing it shouldn't be cleared in the draw function
either. So there's now a shipClearThrusters function that (wait for it)
clears the given ship's activeThrusters object. This must be called
after the ship's update and draw functions.