Stabilizer improvements, thruster clamps, comments
Previously, the stabilizers for each set of thrusters only engaged if
the player wasn't using that set of thrusters. Because the stabilizers
use more thruster strength than the player can, this meant that if the
player was thrusting in one direction and wanted to thrust in the other
the stabilizers would bring them back to neutral faster than they could
bring themselves back. I've changed it so that the stabilizers will be
fired every frame, but will only activate the thruster if the player
either isn't using it or is using it to counteract their current
velocity in that direction. The result is that if the player thrusts
against an existing velocity they'll thrust with their full thruster
strength (the strength the stabilizers use) until they reach 0 velocity
along that axis, at which point they'll continue thrusting with their
typical thruster strength (a third of what the stabilizers use). This
has made the ship much easier to control.

Ships now have thruster clamps, which are limits on their velocity in
each direction of motion. Previously, holding the thruster would cause
the ship to accelerate continuously in that direction without limit.
With thruster clamps, the thrusters will shut off when the ship reaches
the clamp speed, allowing it to only accelerate to a certain point.
Because the stabilizers are now activated every frame, it made sense to
add this code to the stabilizer functions; if the stabilizer for a set
of thrusters detects that the ship's velocity on that thruster's axis
has exceeded the clamp value and the player is still trying to
accelerate in that direction, it shuts off the thruster. Because the
conditions for this conflict with those for stabilization, the ship will
continue to drift at the clamp speed until the player releases the
thruster control for that axis, at which point the stabilizer will
activate normally.

Also, all the code has comments now. Yay comments.