Lasers pt1
Lasers will work mostly the same way as everything else, they'll be
created and added to an array when they're fired, drawn from that array
every frame, and cleared from the array by a clear function that's
called every frame. They'll be described by their start and end vectors
in world space, a color value, and a power value, which represents
damage/stroke width/ etc. So far, I have a createLaser function that
will create a laser with given parameters and add it to the array, and a
shipFireLasers function that determines whether the given ship can fire
a laser, calculates the necessary parameters based on the ship's
properties, and calls the createLaser function. Still to-do is the
function that draws the lasers and the function that does collision
checking.