Lasers pt2, update loop rework
Lasers are now fully functional, though no code has been added to detect
interactions with other objects. But they look nice. The anatomy of a
single laser is as follows:
-lastFireTime: the date in milliseconds the laser was last fired.
Obtained using Date.now()
-cd: the laser's cooldown period in seconds
-range: the length in raw pixels of the laser's beam
-color: obvious
-currentPower: the current strength of the laser. This is set to the max
strength on the frame the laser is fired and is decremented every
subsequent frame by an amount determined by maxPower, coherence, and dt
-coherence: (1-coherence)*maxPower is how much raw power the laser loses
per millisecond. This formula will change, as right now any value below
.9 is essentially an immediate dissipation.
-maxPower: the maximum power value of the laser. This will eventually be
used to determine damage

The way that the laser system works is as follows: At the beginning of a
frame, the array of active lasers is cleared. Input is polled, and any
laser that is being fired this frame has its currentPower value set to
max. During each ship's update loop the currentPower of their lasers are
checked. If the value is more than 0, a laser object is created and
added to the array of active lasers (this object stores start and end
positions in world space, color, and a power value) and the laser's
currentPower is decremented according to its coherence and maxPower.
After all updates have happened collisions are checked, including
collisions with any active laser object. This code doesn't exist yet.
Finally, drawing happens, and any active laser objects are drawn on top
of the grid and under all the other objects. Rinse and repeat.

This is the new structure of the update loop. Frame-specific values
(thruster values, the laser array, etc.) are cleared, actor values are
modified according to input, everything is updated, collisions are
checked, then everything is drawn.