Lasers fully functional (actually this time)
Lasers can now be used to actually destroy asteroids. To describe the
process briefly, the check collision method loops through every active
laser, and for each one iterates through every asteroid, checking the
asteroids distance to the laser's line segment. If the distance is less
than the asteroid's radius and the t value (the distance of the
projection of the point onto the line from the start of the line in
terms of the line's length - so if the closest point on the line is at
75% of its length the t value is .75) is less than the currently
accepted asteroid, it saves the asteroid's object data and the t value.
Once all the asteroids have been evaluated, the chosen asteroid has the
laser's power subtracted from its health, and the t value is used to
shorten the laser so that it ends at the intersection point. Then move
on to the next laser and do the same thing, etc. Asteroids are cleared
with clearAsteroids, which is called at the beginning of update with the
other clear calls, and removes asteroids from the asteroids array if
they have less than 0 hp.