Laser position offset, ship update direction vectors
I've updated updateShip to calculate normalized forward and right
vectors as its first step. It then uses these vectors to apply medial
and lateral acceleration.

I needed to do this because I wanted the laser to start slightly forward
of the ship's center (because the linecap was showing through the back
of the ship) and needed to offset it by a vector along the ship's
direction.