Control mode switching, HUD (sort of), Title/End screens, win condition, star field with parallax
The player can switch the control mode (deactivate the stabilizers) with
tab.

The current control mode will be displayed on-screen at all times.

The game now has a gameState property and a GAME_STATE enum. The game
starts at TITLE, displaying a title screen. It switches to PLAYING when
the player presses W, and proceeds to END when all the asteroids are
destroyed, showing an end screen. No restart yet.

There's a parallaxed star field background now. I generate an array of
stars randomly, with massive radii and world positions, and I create a
star camera, linked to the main camera's canvas context and synced to
the main camera's position and rotation, but with a very low zoom value.
I then draw the stars to that camera, but not to the main one. The
effect is as thought the stars are being drawn by the main camera, but
far in the distance.