otherShips pt1, destructible ships
I've added an array called otherShips that will hold hostile ship
objects. Every ship in this array will have its clearThruster, update,
and draw functions called alongside the player's ship. Each ship in this
array will also have an AI function that gets called alongside the
player's keyboard inputs to determine its thruster values for the frame
and create lasers and all that. This function has a parameter for the
ship to act on and the ship to target. The AI function hasn't been
written yet.

All the ships in otherShips are now subject to collision checking, and
can be destroyed. The player still isn't subject to laser collisions.