Ship AI
I've written a very simple AI function that takes a ship and a target.
It sets the thrusters such that the ship will face the target, maintain
a distance range from it, and orbit such that the target is facing
directly away from it. It will also fire the laser if the target is
within a particular degree spread of its forward vector. It doesn't
calculate optimal thruster percentages or anything like that, so it's
jumpy and inaccurate, but it's insanely agile, and probably still more
accurate than a player. It doesn't track asteroids at the moment though,
it just flies through them, taking damage in the process. I need to fix
that.