Asteroid poly arcing, asteroid/star draw culling
Asteroids are now drawn as a single path and only filled once, rather
than each asteroid having its own path and fill. This resulted in
massive performance increases in any scenario where a large number of
asteroids were visible.

drawAsteroid and drawStar now continue their loops if the current
asteroid or star is entirely off the canvas. Despite the internet
telling me not to bother doing this, it's resulted in much better
scalability with asteroid count and reduced the impact of the star
field. We're talking going from 25 fps with 10000 asteroids to 125 fps.