Camera distance pt1, toggle-able thruster clamps
I figured out how to convert between camera zoom and camera distance,
and set the starfield camera up to always be a certain distance away
from the main camera rather than a static zoom or a fraction of its
zoom. This results in realistic-looking perspective effects between the
playfield and the star field when the player zooms in or out.

Thruster clamps are now toggle-able with C, and tracked in the HUD.