Laser spread, bright colors, hotkey changes
Each laser now has a spread property representing half the maximum
rotational offset in either direction. Bullet spread, basically, but for
lasers.

I've added a method that generates an hsl color with random hue, 100%
saturation, and 75% lightness. This will be used for ship, thruster, and
laser colors instead of getRandomColor, since it looks way nicer.

Rotation is now done with the left and right arrow keys, and lateral
movement with A and D. This is more natural than WASDQE, I think.