Shields, perspective shenanigans, tactical overlay pt1, camera follow
Ships now have shields, with recharge. Shields are a sub-component of
destructible, and have their own constructor function. Recharge happens
in updateShip, because ships are the only things with shields and
because I wasn't sure where else to put it. The shields have nifty
graphics too.

I've added a separate camera for the grid, which will now be drawn 5
units below the gameplay plane. This camera behaves the same as the star
camera, but with a much lower distance offset.

I've also added a sort of tactical overlay which draws information about
the game play objects onto the grid plane. The code is kind of gross -
drawShip and drawAsteroids have been modified to take an optional grid
camera component and draw graphics in its camera space and between its
camera space and the main camera space. I should probably move all the
tactical overlay graphics to their own methods to avoid cluttering the
drawing code for the gameplay objects, but not today.

The main camera has been modified to lerp between its current location
and the player's ship's position plus its velocity vector, giving it the
appearance of leading the ship by a few units. It does this for rotation
as well. I don't personally think it helps the player that much, but
with the camera stuck to the player's ship you lose the perspective
effect on the player's ship, which I think is important for immersion.

This is a big update. The game now looks quite impressive, and with a
few tweaks to the hp values and laser properties I'm actually holding up
pretty well against the AI. This has potential for sure.