Power system pt1
Power system functionality is fully implemented and working. Players can
press one or more of the modifier keys to strengthen their thrusters,
shields, or laser. There are diminishing returns for having more than
one pressed: one gets you the full 1x modifier on the chosen component,
two gives you .5x split across both components, and three gives no
modifier for any component. This system is set up to scale as more
component types are added as well, though I'll need to write an
n-dimensional lerp method for that.

The implementation is similar to the rest of the game, with a few neat
tricks. The powerSystem component itself has two 3D vectors, current and
target, in which each component represents the relative power allotment
for the corresponding component (wording). E.g., [1,1,1] would be equal
power to all components, [1,0,0] would be full power to the first and
none to the rest, [1,1,0] would be half to the first and half to the
second, etc. powerSystem also has transferRate, which is another
lerp-distance per-second thing, used to lerp between current and target
(this architecture seems familiar...). In addition to the powerSystem
component, two functions have been added. scalePowerTarget will take an
n-dimensional vector and scale all its components down or up such that
they sum to 1. getPowerForComponent takes a powerSystem component object
and a component ID from the global SHIP_COMPONENTS enum (which is also
new) and scales the corresponding component in the powerSystem's
n-dimensional current vector such that an original value of (1/n)
becomes 0 and an original value of 1 remains a 1. The current vector is
not scaled in-place; the function clamps the scaled value to 0-1 then
returns it.

Power system targets are cleared along with thrusters or each ship, so
every frame starts with the target at [0,0,0]. The target is populated
at the input check, with each chosen component receiving a target value
of 1. The values are applied in updateShip by dividing all the
augmentable properties by 1 plus the current power value (obtained with
getPowerForComponent) for their component, lerping the powerCystem's
current vector towards its target vector, updating the component power
values with the new current vector, then multiplying the augmentable
properties by 1 plus the power value.

The system was designed this way so that I wouldn't have to re-design
the other components around the power system by storing original and
augmented values for each property I wanted to change. Scaling them
down, updating, then scaling back up allows me to alter any value I want
with the power system regardless of the design of the target component,
and while the transformations up and down aren't 100% accurate, the
values don't seem to drift one way or another over time.

Efficiencies aren't altered for the thrusters or laser, so their
graphics do change slightly in response to power fluctuations, but
shield graphics are entirely unaffected, and the tactical overlay
doesn't reflect power changes in any way. I still have to add that.