Shield efficiency
This is purely a graphical change. The thickness of the shield border
around the ship is now determined by the ship's max value and its
efficiency. A shield whose efficiency matches its max will be 1px thick,
roughly. This means that the size of the shield graphic will increase
with higher power distribution, which is precisely the reason this is
being added.