Camera max zoom, title screen changes
Cameras now have a max zoom as well as a min. 5 seems like a good value
for the main camera. Also, cameras are now created like all the other
components.

The title text has a backdrop now, and the grid will no longer be drawn
on the title screen. Looks pretty good now.