"Efficiency checks"
I tried manual mode a bit and noticed that the thrusters weren't
clearing properly. I had previously wrapped the thruster updates in
updateShip in conditionals so they would only do the thruster math if
the thrusters were actually engaged, but that was before thrusters had
current and target strengths. Because the lerp was happening within that
conditional, the thrusters would only actually move towards their target
if the player, or the stabilizer, was using them, which meant that in
manual mode they wouldn't clear properly. I removed those conditionals,
and now things seem to work as expected.