Mouse controls, some minor refactoring and UI changes
The player's ship can be controlled with the mouse. This was done by
adding a myMouse object to the keys script. The button values are
populated the same way as the keyboard values, but the mouse movement
and mousewheel are done more like the rest of the app: values are
incremented as they come in (via event for these two) and reset to 0
every frame. Unfortunately this means having a resetMouse call in the
update loop, but I can live with that for now.