createShip rework (including thrusterSystem component), HUD redesign pt1, minimap pt1
The thruster portion of createShip has been changed. Previously, ships
had a thrusters object which contained their thruster components, but
wasn't a component itself. Now the thrusters object is a thrusterSystem
component, with its own constructor method.

The HUD has been completely redesigned. All of the information going
forward will be in a slim bar at the bottom of the screen. Clamps and
control system are done, power distribution (called overcharge going
forward) still needs to be done.

We have a sort of working minimap. It doesn't auto-size itself based on
the map size, it isn't positioned properly relative to the main canvas,
and its graphics aren't completed yet, but it's there.