minimap pt2
The minimap works for the current map size now.  I didn't like having
the second canvas, so I scrapped it and used transformations to do the
minimap instead. I had to add a unity-esque viewport component to the
cameras, but only drawMinimap actually uses it. All that I would really
need to add at this point is zoom scaling to get the entire playfield
within the minimap camera's specified viewport, then it would be totally
scalable.