Asteroid collisions
Asteroid collision now behaves as though the ship crashed into the
asteroid and is burrowing through it. Ships will take damage
proportional to their movement speed and be substantially slowed when
colliding with asteroids. The asteroid will now take damage as well. As
of now, AI ships take a tenth as much damage as the player, and the
asteroid takes twice as much as the player, or twenty times as much as
the AI.