Laser damage scaling by distance
Added the T-value damage scaling for lasers, so they do less damage the
farther away their start point is. I'm still using the old formula, so
it isn't 100% accurate, but it's fine. I could even rationalize the
discrepancy away by saying something like "bigger objects have greater
structural density" or something.

(To clarify what I mean - the current formula returns the percentage
from the start of the laser to the end of the laser that's closest to
the center of the colliding object rather than the one that's closest to
the colliding object's edge. So the player will do less damage to
objects with large radii than they probably should)