Overcharge UI, sound/zoom scaling adjustments, add difficulty on win
The overcharge portion of the UI is done. It's pretty minimal.

Thruster and laser sounds will now scale less strongly with zoom. This
is to keep them audible during gameplay.

If the player wins, the max number of ships will double when they reset,
causing the next play to be much harder.