Swept Lasers pt1
Setting the stage for swept area collisions. Laser objects will be
constructed every frame instead of when the laser component's power is
greater than 0, and each laser component will store the previous frame's
laser object. When a laser object is constructed it will contain a
property for the previous frame's laser object, so that when collisions
are detected an area can be swept between the the two laser objects.