Swept area lasers, rendering optimizations
Laser collision now uses a polygon consisting of the start and end of
the current laser object and the object from the previous update, and
SAT. It seems to have made a slight difference, but aiming is still
quite touchy. Damage doesn't currently scale based on the overlap - I
really don't think it needs to.

drawGrid now supports minimap-specific behavior; grid intervals must be
marked with a minimap flag to be drawn on the minimap. This is to allow
only high-interval lines to be drawn on the minimap as an efficiency
measure, rather than every single line.