Projectiles pt1
Ships now have a cannon for firing projectiles. The projectiles will be
spawned with a velocity according to the cannon's power, inheriting the
ship's velocity, and will be culled if they exceed the grid's bounds by
a given percent (30% for now). All that's left to do now is collision
detection