Projectiles pt3
Projectiles do actual damage now. The formula is sort of intricate, it
multiplies the magnitude of the difference in the velocities of the
projectile and the colliding object by the hp of the projectile (hp is
meant to represent mass) and the ratio between the projectile's radius
and the colliding object's radius.

Projectile-asteroid collisions work too, though because of the size
difference the default projectiles are pretty bad against asteroids.
Surprisingly, the collision checks aren't very expensive. Even with
probably near a thousand projectiles and 60 asteroids updates are only
accounting for 5% of total cpu time. Spatial optimization can wait,
thank god.