Ammo, tracer rounds, constructor improvement
Cannons now have an ammo component. This was necessary for the
implementation of tracer rounds, and it opens the door for ammo
switching in the future.

Not every projectile is drawn anymore. Every X number of rounds out of a
cannon will be a tracer round, and only tracer rounds will be drawn.
This is implemented at the ammo level, so if an ammo component has a
tracer interval of 3, every third round out of that ammo component will
be a tracer, remaining consistent across ammo switches, cannon switches,
etc.

Constructors had to be revamped to deal with the ammo implementation.
Previously, if a constructor provided default values for one of its
subcomponents, the calling function couldn't override one of those
values without overriding all of them, as passing an incomplete
definition of the subcomponent to the constructor would cause all of its
default values to be thrown out.

For the new implementation the deepObjectMerge function was added, which
recursively grafts one object onto another. The new constructor pattern
fully initializes the component, using no overrides on any of the
subcomponents, merges the component with an object containing default
values/overrides, then merges the component with the override object
passed to the constructor.

Only the cannon and ammo constructors follow this new pattern so far,
but the others will be refactored in the near future.