Constructor revamp, again
The last commit changed the constructor pattern to use two deep object
merges, first merging default values with the initialized object, then
merging object parameters with the object. I discovered that if any of
the subcomponent constructors have special initialization logic (such as
setting maxHp equal to hp), overrides would obviously break that special
logic.

Instead of doing that, I'm opting to instead deep merge default values
and object parameters on a per-subcomponent basis in the constructor
call for that subcomponent. This allows the overrides to happen before
the special initialization logic.

This made the final deep merge between the new object and the object
parameters overkill, however, but removing that merge would require me
to use the old ternary logic to handle regular (i.e. not for
subcomponents) overrides, which I didn't want to do. So I've added an
extra-shallow merge that completely ignores attributes that are objects.
I'm using that to set regular overrides instead of the deep merge.

Like last time, I haven't changed all the constructors yet. The two I
reworked last time are up to snuff, and createShip is as well, but the
rest are still using the original pattern.