Arbitrary ship updates, updatables pt1
Ship updating is now entirely dynamic. When a ship is created, it will
be sent through populateUpdaters, which will look at the components on
the ship and add any relevant update methods to an array on the ship
object. Ships will be updated by updateUpdatable, which will iterate
through the updaters array and call each function with the ship object
and dT as parameters. This means that ships with any component
configuration are inherently supported, so long as all the components
have update methods and a presence in populateUpdaters.

Creating a ship will now also add it to a new collection, updatables.
The idea is that all component based objects (ships, missiles, drones,
etc) will just be part of this collection, and will be updated
dynamically according to their components. So far I've just created the
collection.