Ship models
Ships now have all their graphics information (vertices, shield
expansion vectors, thruster anchors) saved in a model object, and all
ship graphics will be drawn according to this information.

Vertices represents the vertices in the ship's shape. Pretty self
explanatory.
Shield vectors are the directions each vertex will be pushed in to
generate the shield's shape. These are not normalized.
Thruster points denote the point in model space where each thruster will
be drawn. As of now the thrusters are grouped into medial, lateral, and
rotational, as well as positive and negative, and each direction has
hard-coded direction vectors for drawing the thruster trails. In the
future I might add thruster vectors, but at this point I think it's
fine.

I've also made weapons optional, so they will not be included by
default. If a ship should have weapons, a laser or cannon object will
need to be passed into createShip.

With these changes, everything is finally in place to have different
ship models. So far there are two: cheetah is the original ship shape,
and has a laser, and gull is a diamond shape and has a cannon.