Per-object collision resolution pt2
Laser and ammo components now store a string denoting which collision
function to use. When the laser or projectile object is generated a
function is grabbed from the collisions object using this string and
attached to the laser or projectile object. checkCollisions calls the
attached function when a collision is detected.

The function has to be stored as a string in components so that it can
behave as a primitive type rather than an object, otherwise it might get
mangled by the deep object merge. It's sort of icky, but whatever.