Sweep forwards
All swept volumes are being swept forward in time instead of backwards
now. Lasers were most affected by this. Laser objects from previous
frames will no longer be stored. Instead, checkCollisions will use
createNextLaserObject to predict the position of the next frame's laser
object based on the velocities of its owner.

The impetus for this change was to decouple hitscan object (temporary
lasers) behavior from the firing component in preparation for the
targeting system of the upcoming launcher component, but the removal of
previous frame data from mobile objects is a nice benefit as well.