obj AI rework, tomcat AI started, obj-obj collisions, basic kinetic collision
AI components now contain an AI function, similar to how hitscans and
projectiles contain a collision function, and have an updater. The
updater just calls the AI function with the obj and dt as parameters.

The tomcat missile now has its own AI function. So far all the function
does is activate the medial thruster.

objs can now collide with one another, and they do so using the same
collision method as the default projectile (which has been renamed to
basicKineticCollision). Coincidentally, this function has been reworked.
The base damage taken by each object is the magnitude of the difference
in their velocities (multiplied by dt) multiplied by the other object's
max hp (mass). However, the damage that is actually dealt is clamped to
the ratio between other object's radius, multiplied by a scale factor,
divided by that object's radius. This means that if the other object is
a quarter the size of your ship and the scale factor is 2, the damage
dealt cannot exceed half your ship's health.

To facilitate missiles/rockets, objs needed to be unable to collide with
their owner, but because owner is an object (a JS object) it wouldn't be
picked up by veryShallowObjectMerge when passed into the constructor,
preventing objs from being assigned owners without special logic. In the
interest of not needing to add additional logic every time this sort of
thing happens, createShip's objectParams may now contain an object
called specialProperties that will be added to the ship without any
alteration or cloning.