Targeting system, missile lock
The targeting system is functional. Ships can fire a hitscan to mark
targets, and if the marked target stays within the range and targeting
cone of the targeting system it will be locked, at which point it will
be dropped if it falls out of the targeting system's range. The launcher
will now check for a targeting system and pass a target to the AI of any
launched missiles. The Tomcat AI has very basic tracking as well, and
will vaguely follow its target, detonating within a certain range.

To facilitate passing the target to the AI, a change was made to
deepObjectMerge: any property named specialProperties will not be
recursively merged, and will instead be copied directly from the source
to the target. Without this change, createShip would attempt to deep
clone objectParams.ai before passing it to createComponentAi, which
would both cause the AI to not have updated information about the target
and cause a stack overflow. I'll have to remember going forward that any
component needing access to another obj will need to wrap that obj in a
specialProperties object.