Improved missile AI, overlay configuration, more specialProperties shenanigans
The missile AI is greatly improved, and can now reliably hit stationary
targets. It manages to track ships fairly well, but not well enough to
hit them at the size they are now.

Overlays are now somewhat configurable using the overlay property in
model objects. Options so far include displaying the color circle,
destructible info (shields and health), and range circles for any
components that have ranges (this one is dynamic, any component with a
range property can be added). I'd like to make component-based info as
dynamic as possible, but batching is going to be a big consideration as
things progress, so I might end up hard coding stuff.

Another potential issue with specialProperties was addressed.
Previously, deepObjectMerge and veryShallowObjectMerge would copy the
specialProperties object directly from the target to the source.
However, since the goal was only to ensure that any object references
within specialProperties were were transferred to the target rather than
cloned, it is safer to shallow-clone the specialProperties object before
adding it to the target. This ensures that the only reference to the
specialProperties object in the target is the reference from the target,
preventing external mutation. This may or may not have been the cause of
a missile glitch; I didn't investigate before making the change, and it
hasn't happened since. I'm hoping this will be the last time I touch
specialProperties.