Respawns
objs with a respawnTime property on their objectParams will be give
queueRespawn as a destructor, which will add them to the respawnQueue
for their game. update will iterate through the queue every update, and
if any of its elements are ready for respawn it will pass their params
property to createShip.

Because the player's ship is a special object, it doesn't respawn at the
moment.