client-server architecture pt2
Camera creation and manipulation, input gathering, and state tracking
have been moved to the client. The server contains a list of running
game objects, and as soon as a game object is created its simulation
starts regardless of connection status.

The client and server communicate via a simulated websocket. This gets
created at client start, after the main game has been created and
started on the server. When the connection is created, the server
creates a player object for it containing the socket, the game the
player is currently in, and the ship the player is currently flying.
Both of these will be undefined initially.

When the palyer presses enter at the start screen, the client enters the
WAIT state and sends a message to the server containing the ID of the
game to join and the name (string) of the ship to fly. On receiving this
message the server adds the player to the given game if it exists, then
creates their desired ship if it exists and adds it to the game.

While the game is playing, all connected players will have their ship's
position/orientation and velocity sent through their socket (along with
the game object, for now) every update (will eventually be less
frequently). The client will use this information in its updates to
control camera movement.

Inputs from the player are sent through the socket. At the moment key up
and key down events are sent as soon as they occur, but eventually all
inputs will be batched and sent all at once at some interval. When they
arrive at the server they are passed to the player's ship's remoteInput
object, whose update function uses them to control the ship (this update
function replaces objKeyboardControl).

Because I'm not sending HUD-related information back, the HUD still
doesn't work, and obviously I need to divorce drawing from the actual
game object, but this is a great start.