Full ship controls working, player-ship communication rework
All the ship controls are now working. For one-shot inputs the
remoteInput will check the state of the key, and if it's down it will do
the action then set it to up.

Communication between the player object on the server and that player's
ship has been reworked. Instead of holding a reference to the ship obect
itself, the palyer object will have a remoteSend property containing the
remoteInput's message handler, and the remoteInput will have a
remoteSend property containing an anonymous function that calls the send
function of the player's socket. All communication between the client
and the player's ship will be done through these remoteSend functions.

I've also added some code to hopefully prevent memory leaks on socket
closes and player ship destructions. If the socket is closed, the server
will send a disconnect message through that player's remoteSend that
will cause the remoteInput to delete its remoteSend. Ships containing a
remoteInput will be given a destructor that sends a destruction message
through its remoteSend which will cause the player object to delete its
remoteSend. This will, hopefully, allow the disconnected player object
or the destroyed ship to fall totally out of scope and be garbage
collected.