World info drawing pt1
Info about objs, asteroids, projectiles, hitscans, and radials is now
sent via the remoteInput and used for drawing. As of now the only thing
drawn directly from the game object is the grid. Ship overlays aren't in
yet, I need to rewrite them a bit to get them to work with the new
architecture.