Camera jitter
So the jitter I was seeing in ship movement was actually coming from two
places. The first is temporal aliasing between the client and the
server, which I knew about. The second was aliasing between the client's
update and draw loops, which was occurring because I was, for whatever
reason, trying to update the camera in the update loop rather than the
draw loop. Fixing that cut the jitter substantially, but until I can
eliminate it completely I'm going to get rid of the lerping and just put
the camera at the player's location.