Interpolated graphics
All the information sent to the client is interpolated for drawing. A
metric fuck load of changes were required for this, the most important
being an ID system for basically every object on the server and a maor
restructuring of worldInfo on the client.

Each obj, projectile, hitscan*, radial, and obj component created on the
server will be issued a unique integer ID when created using
server.takeIdTag. Basically, the server gives out IDs using an
incrementing int and logs them in a dictionary. When objects are
destroyed their ID is removed from the dictionary so it can be used
again. takeIdTag will loop when it reaches the max safe int (some
ridiculous number, probably won't ever happen), and as it gives out IDs
it checks that they aren't in use first.

worldInfo now contains an array for every object type as well as three
new objects, drawing, targets, and previousTargets. The life cycle for
an object coming in over the socket is as follows:
-It comes in as an array on data.worldInfo and gets passed to
pushCollectionFromDataToWi.
-pushCollection sees that its ID isn't in drawing, so it sets
drawing[ID] to false, copies the object into previousTargets[ID], and
pushes the object to its typed array.
-Drawing code sees that the object'd ID is set as false in drawing and
skips it.
-The next set of info for the object (meaning a new object with the same
ID) comes over the socket. Everything in targets gets copied to
previousTargets with the same keys, and targets is cleared. The object
gets passed to pushCollection, which, seeing that the ID is in drawing,
copies it to targets[ID] and sets drawing[ID] to true.
-Because drawing is now true for the object, the drawing code doesn't
skip it. Any properties on the object that are used by drawing code are
first updated with interpolateWiValue, which uses previousTarget,
target, the timestamp of the last worldInfo message, and the message
interval (sent by the player's ship upon creation, since the interval is
controlled by the remoteInput) to lerp the value to the proper point.
-This continues until either the object is destroyed on the server or
the player gets far enough away from it to not receive its info in world
updates. Because targets is cleared before every world update our
object's Id will no longer be in the targets object.
-Drawing code sees that the object has no target and uses
removeIndexFromWiCollection to remove the object from targets,
previousTargets, drawing, and its typed collection. The drawing code
decrements its counter and continues on with whatever it was doing.

Pretty spiffy. Unfortunately it also makes the client code kind of a
headache.

There are two bugs at the moment:
-Because the camera uses the non-lerped player position it's still
jittery. I'll have to set up a lerping procedure for the player position
so it matches where the player's ship actually is. More headaches.
-Because remoteInput stamps and sends its objects before collisions
happen lasers appear to pass through things. remoteInput is going to
have to queue its sends so that they can be processed after collisions.
More headaches...

This is fantastic progress though.